Descriptive stats added.
The code puts graphs in the working directory NOT in the graphs folder - why?

It looks like some graphs and tables are duplicative. Please rectify.

When you make a graph, it is a good idea to double check the output. Many of the titles don't fit, you have to adjust the output with \n to make the title wrap or use a more sophistocated approach.